+++ app/models/conversion/task_list.rb
@@ -11,7 +11,8 @@ class Conversion::TaskList < ::BaseTaskList
           Conversion::Task::ConversionGrantTaskForm,
           Conversion::Task::SponsoredSupportGrantTaskForm,
           Conversion::Task::AcademyDetailsTaskForm,
-          Conversion::Task::MainContactTaskForm
+          Conversion::Task::MainContactTaskForm,
+          Conversion::Task::ProposedCapacityOfTheAcademyTaskForm
         ]
       },
       {