+++ spec/models/conversion/task_list_spec.rb
@@ -14,6 +14,7 @@ RSpec.describe Conversion::TaskList do
         :sponsored_support_grant,
         :academy_details,
         :main_contact,
+        :proposed_capacity_of_the_academy,
         :land_questionnaire,
         :land_registry,
         :supplemental_funding_agreement,
@@ -55,7 +56,8 @@ RSpec.describe Conversion::TaskList do
               Conversion::Task::ConversionGrantTaskForm,
               Conversion::Task::SponsoredSupportGrantTaskForm,
               Conversion::Task::AcademyDetailsTaskForm,
-              Conversion::Task::MainContactTaskForm
+              Conversion::Task::MainContactTaskForm,
+              Conversion::Task::ProposedCapacityOfTheAcademyTaskForm
             ]
           },
           {
@@ -116,7 +118,7 @@ RSpec.describe Conversion::TaskList do
       project = create(:conversion_project)
       task_list = described_class.new(project, user)
 
-      expect(task_list.tasks.count).to eql 29
+      expect(task_list.tasks.count).to eql 30
       expect(task_list.tasks.first).to be_a Conversion::Task::HandoverTaskForm
       expect(task_list.tasks.last).to be_a Conversion::Task::ReceiveGrantPaymentCertificateTaskForm
     end